+++ mobile/.gitignore
@@ -1,4 +1,3 @@
 # Except this file
 !.gitignore
-!gradle-wrapper.jar
 !store
