+++ services/analytics/src/main/java/com/wedriveu/services/analytics/entity/AnalyticsStore.java
@@ -1,13 +1,13 @@
 package com.wedriveu.services.analytics.entity;
 
-import com.wedriveu.services.shared.entity.AnalyticsVehicle;
-import com.wedriveu.services.shared.entity.VehicleCounter;
+import com.wedriveu.services.shared.model.AnalyticsVehicle;
+import com.wedriveu.services.shared.model.VehicleCounter;
 
 import java.util.Optional;
 
 /**
  * @author Stefano Bernagozzi
- * This inteface models the <em>com.wedriveu.services.shared.entity.AnalyticsVehicle' database store</em>.
+ * This inteface models the <em>AnalyticsVehicle' database store</em>.
  */
