+++ services/analytics/src/main/java/com/wedriveu/services/analytics/entity/AnalyticsStoreImpl.java
@@ -1,8 +1,8 @@
 package com.wedriveu.services.analytics.entity;
 
-import com.wedriveu.services.shared.entity.AnalyticsVehicle;
-import com.wedriveu.services.shared.entity.EntityListStoreStrategy;
-import com.wedriveu.services.shared.entity.VehicleCounter;
+import com.wedriveu.services.shared.model.AnalyticsVehicle;
+import com.wedriveu.services.shared.store.EntityListStoreStrategy;
+import com.wedriveu.services.shared.model.VehicleCounter;
 import com.wedriveu.services.shared.util.Log;
 
 import java.util.ArrayList;
