+++ services/analytics/src/main/scala/com/wedriveu/services/analytics/entity/VehiclesCounterAlgorithm.scala
@@ -1,6 +1,7 @@
 package com.wedriveu.services.analytics.entity
 
-import com.wedriveu.services.shared.entity.{AnalyticsVehicle, VehicleCounter}
+import com.wedriveu.services.shared.entity.AnalyticsVehicle
+import com.wedriveu.services.shared.model.{AnalyticsVehicle, VehicleCounter}
 
 /**
   * @author Stefano Bernagozzi
