+++ services/analytics/src/main/scala/com/wedriveu/services/analytics/entity/VehiclesCounterAlgorithmImpl.scala
@@ -1,7 +1,8 @@
 package com.wedriveu.services.analytics.entity
 
 
-import com.wedriveu.services.shared.entity.{AnalyticsVehicle, VehicleCounter}
+import com.wedriveu.services.shared.entity.AnalyticsVehicle
+import com.wedriveu.services.shared.model.{AnalyticsVehicle, VehicleCounter}
 
 import scala.collection.JavaConverters._
