+++ services/analytics/src/test/java/com/wedriveu/services/analytics/entity/AnalyticsStoreImplTest.java
@@ -1,10 +1,10 @@
 package com.wedriveu.services.analytics.entity;
 
 
-import com.wedriveu.services.shared.entity.AnalyticsVehicle;
-import com.wedriveu.services.shared.entity.EntityListStoreStrategy;
-import com.wedriveu.services.shared.entity.JsonFileEntityListStoreStrategyImpl;
-import com.wedriveu.services.shared.entity.VehicleCounter;
+import com.wedriveu.services.shared.model.AnalyticsVehicle;
+import com.wedriveu.services.shared.store.EntityListStoreStrategy;
+import com.wedriveu.services.shared.store.JsonFileEntityListStoreStrategyImpl;
+import com.wedriveu.services.shared.model.VehicleCounter;
 import org.junit.Before;
 import org.junit.Test;
