+++ services/authentication/src/main/java/com/wedriveu/services/authentication/control/CredentialsCheckerImpl.java
@@ -2,7 +2,7 @@ package com.wedriveu.services.authentication.control;
 
 import com.wedriveu.services.authentication.entity.UserStore;
 import com.wedriveu.services.authentication.entity.UserStoreImpl;
-import com.wedriveu.services.shared.entity.User;
+import com.wedriveu.services.shared.model.User;
 import com.wedriveu.services.shared.util.Log;
 
 import java.io.IOException;
