+++ services/authentication/src/main/java/com/wedriveu/services/authentication/entity/UserStoreImpl.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.core.JsonGenerationException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.JsonMappingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.wedriveu.services.shared.entity.User;
+import com.wedriveu.services.shared.model.User;
 import com.wedriveu.services.shared.util.Log;
 
 import java.io.File;
