+++ services/booking/src/main/java/com/wedriveu/services/booking/entity/BookingStoreImpl.java
@@ -1,7 +1,7 @@
 package com.wedriveu.services.booking.entity;
 
-import com.wedriveu.services.shared.entity.Booking;
-import com.wedriveu.services.shared.entity.EntityListStoreStrategy;
+import com.wedriveu.services.shared.model.Booking;
+import com.wedriveu.services.shared.store.EntityListStoreStrategy;
 import com.wedriveu.services.shared.util.Log;
 
 import java.util.ArrayList;
