+++ services/booking/src/test/java/com/wedriveu/services/booking/entity/BookingStoreTest.java
@@ -1,8 +1,8 @@
 package com.wedriveu.services.booking.entity;
 
-import com.wedriveu.services.shared.entity.Booking;
-import com.wedriveu.services.shared.entity.EntityListStoreStrategy;
-import com.wedriveu.services.shared.entity.JsonFileEntityListStoreStrategyImpl;
+import com.wedriveu.services.shared.model.Booking;
+import com.wedriveu.services.shared.store.EntityListStoreStrategy;
+import com.wedriveu.services.shared.store.JsonFileEntityListStoreStrategyImpl;
 import com.wedriveu.shared.util.Position;
 import org.junit.Before;
 import org.junit.Test;
