+++ services/shared/.gitignore
@@ -1,0 +1,3 @@
+# Except this file
+!.gitignore
+!src/main/java/com/wedriveu/services/shared/store
