+++ services/shared/src/main/java/com/wedriveu/services/shared/message/VehicleResponseCanDrive.java
@@ -1,4 +1,4 @@
-package com.wedriveu.services.shared.rabbitmq.nearest;
+package com.wedriveu.services.shared.message;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
