+++ services/shared/src/main/java/com/wedriveu/services/shared/model/VehicleCounter.java
@@ -1,4 +1,4 @@
-package com.wedriveu.services.shared.entity;
+package com.wedriveu.services.shared.model;
 
 /**
  * @author Stefano Bernagozzi
