+++ services/shared/src/main/java/com/wedriveu/services/shared/rabbitmq/ExchangeManagerVerticle.java
@@ -1,5 +1,6 @@
 package com.wedriveu.services.shared.rabbitmq;
 
+import com.wedriveu.services.shared.message.SharedMessages;
 import com.wedriveu.services.shared.rabbitmq.client.RabbitMQClientFactory;
 import com.wedriveu.services.shared.util.Log;
 import io.vertx.core.AbstractVerticle;
@@ -33,7 +34,7 @@ public class ExchangeManagerVerticle extends AbstractVerticle {
     }
 
     private void declareExchange(String exchangeName) {
-        client.exchangeDeclare(exchangeName,DIRECT, false, false, onDeclareCompleted -> {
+        client.exchangeDeclare(exchangeName, DIRECT, true, false, onDeclareCompleted -> {
             if (onDeclareCompleted.succeeded()) {
                 vertx.eventBus().send(SharedMessages.VehicleService.EXCHANGE_BINDED, null);
             } else {
