+++ services/shared/src/test/java/com/wedriveu/services/shared/entity/EntityListStoreStrategyTest.java
@@ -1,5 +1,7 @@
 package com.wedriveu.services.shared.entity;
 
+import com.wedriveu.services.shared.store.EntityListStoreStrategy;
+import com.wedriveu.services.shared.store.JsonFileEntityListStoreStrategyImpl;
 import org.junit.Before;
 import org.junit.Test;
