+++ services/vehicle/src/main/java/com/wedriveu/services/vehicle/boundary/nearest/VehicleElectionVerticle.java
@@ -1,6 +1,6 @@
 package com.wedriveu.services.vehicle.boundary.nearest;
 
-import com.wedriveu.services.shared.entity.Vehicle;
+import com.wedriveu.services.shared.model.Vehicle;
 import com.wedriveu.services.shared.rabbitmq.VerticlePublisher;
 import com.wedriveu.services.shared.util.Log;
 import com.wedriveu.services.vehicle.rabbitmq.Messages;
