+++ services/vehicle/src/main/java/com/wedriveu/services/vehicle/rabbitmq/UserRequest.java
@@ -1,7 +1,7 @@
 package com.wedriveu.services.vehicle.rabbitmq;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.wedriveu.services.shared.entity.Vehicle;
+import com.wedriveu.services.shared.model.Vehicle;
 import com.wedriveu.shared.util.Position;
 
 import java.util.List;
