+++ services/vehicle/src/test/java/com/wedriveu/services/vehicle/boundary/nearest/VehicleNearestVerticleTest.java
@@ -1,6 +1,6 @@
 package com.wedriveu.services.vehicle.boundary.nearest;
 
-import com.wedriveu.services.shared.entity.Vehicle;
+import com.wedriveu.services.shared.model.Vehicle;
 import com.wedriveu.services.vehicle.app.BootVerticle;
 import com.wedriveu.services.vehicle.boundary.BaseInteractionClient;
 import com.wedriveu.services.vehicle.boundary.nearest.entity.UserDataFactoryA;
