+++ services/vehicle/src/test/java/com/wedriveu/services/vehicle/boundary/vehicleregister/entity/VehicleFactoryMini.java
@@ -1,6 +1,6 @@
 package com.wedriveu.services.vehicle.boundary.vehicleregister.entity;
 
-import com.wedriveu.services.shared.entity.Vehicle;
+import com.wedriveu.services.shared.model.Vehicle;
 import com.wedriveu.shared.util.Position;
 
 import java.net.MalformedURLException;
